setTimeout(greeting. Changed to to 500. Message appear faster. 500 milliseconds. Tried 3000 as well. Slower apperance on message.
added consol log. Nothing appears on the site. BUT when I add console log under: if (sender === 'bot ') {
The message appears in the console. It’s because user hasn’t been pushed to form yet. But the bot has due to his welcome message. I guess!

When you remove dollar sign and curly brackets the word “message” appears in the chat. (bot not user). The ${ } seems to work like a placeholder?